Applied created appropriate packages and moved appropriate elements inside of them, following app architecture
