+++ app/src/main/java/co/develhope/meteoapp/data/domain/WeatherIcon.kt
@@ -1,4 +1,4 @@
-package co.develhope.meteoapp.today
+package co.develhope.meteoapp.data.domain
 
 import co.develhope.meteoapp.R
