+++ app/src/main/java/co/develhope/meteoapp/data/domain/setImageResource.kt
@@ -1,4 +1,4 @@
-package co.develhope.meteoapp.today
+package co.develhope.meteoapp.data.domain
 
 import android.widget.ImageView
